chatbot/requirements.txt
langgraph==0.2.0
langchain==0.3.0
langchain-core==0.3.29
langchain-community==0.3.0
deepseek-sdk==0.1.0
sqlalchemy==2.0.23
python-dotenv==1.0.0